test: animations - ⭕ IconAnimator
